# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: patchwright
  version: v1
  description: |
    Read-only access to the latest cached assessment, plus ticket reconciliation.

    Every endpoint except `/healthz`, `/readyz` and `/metrics` requires the shared
    token when `PATCHWRIGHT_API_TOKEN` is set: `Authorization: Bearer <token>`, or
    HTTP Basic with the token as the password so a browser can be prompted.

    Absent data and zero are different states throughout. A finding with
    `provider_assessed: false` has zero severity counts because nothing looked, not
    because nothing was found; `scanned` and `exploit_checked` say the same for the
    optional scanners. Consumers that treat them as equivalent will report an
    unscanned estate as a clean one.
  license:
    name: MIT
    url: https://github.com/s-humphreys/patchwright/blob/main/LICENSE
servers:
  - url: http://localhost:8080
    description: Local or port-forwarded
security:
  - bearerAuth: []
...
paths:
...
  /tickets:
    get:
      tags: [operations]
      summary: Ticket plan page
      description: |
        HTML, and a page of its own rather than a tab on the dashboard: the dashboard's
        subject is the estate, while this is what reconciliation would write to a
        tracker. Reads GET /api/v1/tickets. Read-only - opening it applies nothing.
      responses:
        "200":
          description: The ticket plan page
          content:
            text/html:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /analytics:
    get:
      tags: [operations]
      summary: Security analytics page
      description: |
        HTML. Per-team responsiveness: who is working through their findings and who is
        sitting on them. A page of its own because its subject is how teams are
        RESPONDING, not what is wrong - a different question read by different people.
        Reads GET /api/v1/analytics.
      responses:
        "200":
          description: The analytics page
          content:
            text/html:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/analytics:
    get:
      tags: [analysis]
      summary: Per-team responsiveness
      description: |
        Who is moving on their findings and who is not, computed from the latest cached
        assessment.

        The headline metric is `stale_unstarted`: actionable findings with an upgrade
        available, no open pull request, no ticket, and older than `stale_fix_days`. A
        fix has existed for a month and nothing has moved, which is a statement about
        the team rather than about the software.

        `notes` lists the questions this endpoint is asked and cannot answer - read it
        before building a dashboard on this, because the gaps are the part most likely
        to be assumed away. Ticket resolution time is the notable one: the tracker index
        holds only open tickets and carries no dates.
      responses:
        "200":
          description: Per-team responsiveness
          content:
            application/json:
              schema:
                type: object
                required: [assessment, analytics]
                properties:
                  assessment: { $ref: "#/components/schemas/AssessmentMeta" }
                  analytics: { $ref: "#/components/schemas/Analytics" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/findings:
    get:
      tags: [findings]
      summary: List findings
      parameters:
        - { name: owner_class, in: query, schema: { type: string }, description: "Owner class, e.g. platform" }
        - { name: team, in: query, schema: { type: string } }
        - { name: repository, in: query, schema: { type: string }, description: "Bare repository, e.g. storefront" }
        - { name: priority, in: query, schema: { type: string, enum: [urgent, high, medium, low] } }
        - { name: actionable, in: query, schema: { type: boolean } }
        - { name: live, in: query, schema: { type: boolean } }
        - { name: upgradable, in: query, schema: { type: boolean } }
        - { name: known_exploited, in: query, schema: { type: boolean } }
        - { name: suppressed, in: query, schema: { type: boolean } }
        - name: provider_assessed
          in: query
          schema: { type: boolean }
          description: "`false` lists the coverage gap: findings nobody has data for."
        - { name: remediation_checked, in: query, schema: { type: boolean } }
        - { name: upgrade_resolved, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Matching findings
          content:
            application/json:
              schema:
                type: object
                required: [assessment, findings]
                properties:
                  assessment: { $ref: "#/components/schemas/AssessmentMeta" }
                  findings:
                    type: array
                    items: { $ref: "#/components/schemas/Finding" }
                  tickets:
                    description: |
                      Open tickets keyed by image repository. Absent when Jira is not
                      configured, which means unknown rather than "no ticket exists".
                    type: object
                    additionalProperties:
                      type: array
                      items: { $ref: "#/components/schemas/TicketRef" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/NoAssessment" }
  /api/v1/finding:
    get:
      tags: [findings]
      summary: One image's finding
      parameters:
        - name: image
          in: query
          required: true
          schema: { type: string }
          description: Full image reference or repository
      responses:
        "200":
          description: The finding
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Finding" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: No finding for that image }
  /api/v1/items:
    get:
      tags: [findings]
      summary: The queue as work items
      description: |
        Findings grouped into the units people act on: one item per owning team, service
        and upgrade target — the same key a ticket uses, so an item, a queue row and a
        ticket are one thing.

        Every aggregate states its own limits. `priority` is the worst across the item's
        deployments and `priority_where` names where that came from; `assessed_images`
        short of `deployments` means the counts are the worst KNOWN rather than the
        worst; `in_flight_checked` is true only when every deployment was checked.

        Accepts the same filters as /api/v1/findings.
      parameters:
        - { name: team, in: query, schema: { type: string } }
        - { name: owner_class, in: query, schema: { type: string } }
        - { name: repository, in: query, schema: { type: string } }
        - { name: priority, in: query, schema: { type: string, enum: [urgent, high, medium, low] } }
        - { name: upgradable, in: query, schema: { type: boolean } }
        - { name: known_exploited, in: query, schema: { type: boolean } }
        - { name: suppressed, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Work items, worst first
          content:
            application/json:
              schema:
                type: object
                required: [assessment, count, items]
                properties:
                  assessment: { $ref: "#/components/schemas/AssessmentMeta" }
                  count: { type: integer }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Item" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/service:
    get:
      tags: [findings]
      summary: One service's outstanding work
      description: |
        The work items for a single repository — one request for a service catalogue
        component page, keyed by the repository the catalogue already knows.

        `known` is false when this assessment contains no finding for the service at
        all. An empty `items` list with `known: false` is ignorance rather than health,
        and a consumer cannot tell the two apart from the array alone.
      parameters:
        - { name: repository, in: query, required: true, schema: { type: string } }
        - { name: team, in: query, schema: { type: string } }
        - { name: suppressed, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: The service's items
          content:
            application/json:
              schema:
                type: object
                required: [assessment, repository, known, count, items]
                properties:
                  assessment: { $ref: "#/components/schemas/AssessmentMeta" }
                  repository: { type: string }
                  known: { type: boolean }
                  count: { type: integer }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Item" }
        "400": { description: repository is required }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/NoAssessment" }
  /api/v1/cves:
    get:
      tags: [findings]
      summary: The estate by CVE
      description: |
        Vulnerabilities across the estate, worst first: known-exploited, then severity,
        then exploitation pressure, then reach. This is the scope-of-work question,
        which the per-service queue can only answer by reading every row.

        `images` and `services` are both reported because they answer different
        questions: how many deployments carry it, and how many pieces of work fixing it
        takes. One base-image CVE on 535 images is a handful of rebuilds.

        Per-CVE detail exists only where a vulnerability source ran, so
        `scanned_findings` and `total_findings` travel with the answer. Zero CVEs over
        zero scanned findings means nothing was looked at, not that nothing was found.
      parameters:
        - { name: severity, in: query, schema: { type: string, enum: [critical, high, medium, low] } }
        - { name: kev, in: query, schema: { type: boolean }, description: CISA known-exploited only }
        - name: fixable
          in: query
          schema: { type: boolean }
          description: |
            true keeps CVEs with a fix on at least one affected image; false keeps those
            with none anywhere, which is the waiting-on-upstream list.
        - { name: min_images, in: query, schema: { type: integer } }
        - { name: min_services, in: query, schema: { type: integer } }
        - { name: team, in: query, schema: { type: string } }
      responses:
        "200":
          description: CVEs, worst first
          content:
            application/json:
              schema:
                type: object
                required: [assessment, count, scanned_findings, total_findings, cves]
                properties:
                  assessment: { $ref: "#/components/schemas/AssessmentMeta" }
                  count: { type: integer }
                  scanned_findings: { type: integer }
                  total_findings: { type: integer }
                  cves:
                    type: array
                    items: { $ref: "#/components/schemas/CVE" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/cve:
    get:
      tags: [findings]
      summary: One CVE and every image carrying it
      parameters:
        - { name: id, in: query, required: true, schema: { type: string } }
        - { name: team, in: query, schema: { type: string } }
      responses:
        "200":
          description: The CVE, with its affected images
          content:
            application/json:
              schema:
                type: object
                required: [assessment, cve]
                properties:
                  assessment: { $ref: "#/components/schemas/AssessmentMeta" }
                  cve: { $ref: "#/components/schemas/CVE" }
        "400": { description: id is required }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: |
            No scanned image carries it. The response carries scanned_findings and
            total_findings, because "not found" over an unscanned estate means nothing
            looked rather than nothing carries it.
        "503": { $ref: "#/components/responses/NoAssessment" }
  /api/v1/owners:
    get:
      tags: [findings]
      summary: Per-team triage
      responses:
        "200":
          description: One row per owner class and team
          content:
            application/json:
              schema:
                type: object
                required: [assessment, owners]
                properties:
                  assessment: { $ref: "#/components/schemas/AssessmentMeta" }
                  owners:
                    type: array
                    items: { $ref: "#/components/schemas/OwnerStats" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/NoAssessment" }
  /api/v1/summary:
    get:
      tags: [findings]
      summary: Fleet-wide headline and coverage
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                type: object
                required: [assessment, summary]
                properties:
                  assessment: { $ref: "#/components/schemas/AssessmentMeta" }
                  summary: { $ref: "#/components/schemas/Summary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/NoAssessment" }
  /api/v1/config:
    get:
      tags: [operations]
      summary: The loaded rules
      description: |
        The ownership and policy configuration as parsed at startup — not the files as
        they are on disk now, which may include edits that are not in effect.

        Values of keys that look like credentials are replaced with `[redacted]`.
        Nothing should ever match, since credentials come from the environment by
        design; the check exists because the cost of being wrong is a credential in a
        browser tab.
      responses:
        "200":
          description: One entry per config file, in load order
          content:
            application/json:
              schema:
                type: object
                required: [sources]
                properties:
                  sources:
                    type: array
                    items:
                      type: object
                      required: [path, content]
                      properties:
                        path: { type: string }
                        content: { type: string, description: YAML as loaded. }
                        redacted:
                          type: boolean
                          description: A value in this file was withheld.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: No configuration was loaded }
  /api/v1/assessments:
    post:
      tags: [operations]
      summary: Trigger a refresh
      description: Returns immediately; the refresh runs in the background.
      responses:
        "202":
          description: Refresh started, or already running
          content:
            application/json:
              schema:
                type: object
                properties:
                  assessment: { $ref: "#/components/schemas/AssessmentMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /static/app/:
    get:
      tags: [operations]
      summary: Status page assets
      description: |
        The page's JavaScript modules and stylesheet, appended to this prefix (e.g.
        `/static/app/main.js`), from the same embedded tree as the page. Behind the same
        auth as the page, since together they are the page.
      responses:
        "200":
          description: The asset
          content:
            text/javascript:
              schema: { type: string }
            text/css:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: No such asset }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `PATCHWRIGHT_API_TOKEN`. HTTP Basic with any username and the token as the
        password also works, so a browser can be prompted for the status page.
  responses:
    Unauthorized:
      description: Missing or invalid token
    BadRequest:
      description: Invalid parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    NoAssessment:
      description: No assessment has completed yet
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
  schemas:
...
    TicketRef:
      type: object
      properties:
        key: { type: string }
        status: { type: string }
        url: { type: string }
...